fix: address remaining P2s from cross-cutting audit (#435)
Follow-ups to the previous commit that close out the rest of the audit
findings rather than punting them to a separate PR:

- log/handler.go: multiHandler.Handle aggregates errors via errors.Join
  so a failure in one child no longer skips later handlers.
- log/logger.go: document that Fatal goes straight to os.Exit and won't
  flush deferred Sync hooks on buffered or async handlers — relevant if
  callers add a file-rotation sink behind bufio.Writer.
- internal/observability: move resetForTest, samplerDoneForTest, and
  samplesCountForTest into export_test.go so they no longer ship in the
  production binary; drop the unused SetSampleForTest helper.
- internal/testing/env_pseudo.go: tag EnableFeature/DisableFeature as
  testing helpers and document the Close()-required semantic mirrored
  from rippled's Env::enableFeature.
